app/requirements.txt
opencv-python
imageio
imageio-ffmpeg
ffmpeg-python
av
runpod==1.6.0
xformers==0.0.25
decord==0.6.0
einops==0.3.0
numpy==1.24.2
omegaconf==2.1.1
opencv_python
pandas==2.0.0
Pillow==9.5.0
pytorch_lightning==1.9.3
PyYAML==6.0
setuptools==65.6.3
tqdm==4.65.0
transformers==4.25.1
moviepy
timm
scikit-learn
open_clip_torch==2.22.0
kornia
diffusers
gradio
boto3
